mvc/testTools/requirements.txt
Django==1.9.1
sqlite-devel
djangorestframework
